chore: move PrometheusJsonResponse to a separate file and impl IntoResponse
